Add option --quiet and refactor human/raw setting
The option quiet remove the output of lines ON-...

It is needed to remove outputs that are useful to humans
but not to programs

Conversely, the opposition human/raw is now more simple and
is then simplified in the code.

At the same time, the output of direct reply is changed, the emulation
of the structured response is removed.